feat(client): 프로필 카드 클릭 로깅 추가
